dom/base/test/test_meta_viewport_device_width_with_initial_scale_2.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>device-width with initial-scale=2 in meta viewport</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta name="viewport" content="width=device-width, initial-scale=2">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <p>width=device-width, initial-scale=2</p>
  <script type="application/javascript">
    "use strict";

    add_task(async function device_width_with_initial_scale_2() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      // In the case device is specified, max-width isn't 'extend-to-zoom', it's
      // the display width, and if the initial-scale is greater than 1,
      // min-width will be the same as max-width because extend-width will be
      // less than the display width, thus the final width should be the display
      // width.
      is(info.width, 800, "width should be the display width");
      is(info.height, 480, "height should be the display height");
      is(info.defaultZoom, 2, "initial-scale should be 2");
    });
  </script>
</body>
</html>
